requirement.txt
pillow==10.2.0
pyfiglet==1.0.2
base64